posix: declare timespec_to_timeoutms() in posix_clock.h
Provide a single declaration of timespec_to_timeoutms() (which is
a private function), in the private header file posix_clock.h .

+++ lib/posix/options/timespec_to_timeout.c
@@ -4,11 +4,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+#include "posix_clock.h"
+
 #include <zephyr/kernel.h>
 #include <ksched.h>
 #include <zephyr/posix/time.h>
 
-int64_t timespec_to_timeoutms(const struct timespec *abstime)
+uint32_t timespec_to_timeoutms(const struct timespec *abstime)
 {
 	int64_t milli_secs, secs, nsecs;
 	struct timespec curtime;
